test: split Claude reset confirmation coverage
